Generalization decoding + Sam's data processing + ML eventmarker overwrites + rewardAnticipatory subEvent
Major changes in runAnalyses are related to getting Sam's data to go in smoothly (new checks for 'Sam' or 'Mo' in filename).

One of those changes was making a system which could overwrite the wrong codes on the blackrock system w/ properly timed codes in ML files. Most of this is in 'replacePacketData'.

First update since May 25th:
1. error for buildRunList is now outside, rather than inside.
2. couple new utility functions like 'binStuff', 'indicies2Matrix', 'saveOpenFigs',
3. Cleaned up plot_per_label_accuracy a bit
4. createBarPlotWithChanceLine updated to not occlude any bars w/ legend.
5. 'exampleCellFinderObject' developed.
6. Made necessary updates to code which normalizes PSTHes in batchAnalysis, specifically using FixAligned versions of psthes to calculate baseline, along with pulling the right parameters to find that.

More broadly, implemented the 'generalization decoding' - which trains and tests on different instances of a stimulus.

Also added anticipatory reward event in subEventAnalysis script.

- Next Steps:
- Integrate PCA_test/PCA_vis_test into the rest of the batchAnalysis scripts, and follow up with testing.
- saccade per subEvent